+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/SettingsTests.cs
@@ -351,6 +351,7 @@ public class SettingsTests : IDisposable
 
     [Theory]
     [InlineData("CONTAINER", ResourceDetector.Container)]
+    [InlineData("AZURE", ResourceDetector.Azure)]
     internal void GeneralSettings_Instrumentations_SupportedValues(string resourceDetector, ResourceDetector expectedResourceDetector)
     {
         Environment.SetEnvironmentVariable(ConfigurationKeys.ResourceDetectorEnabled, "false");